requirements.txt
# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 向量数据库
qdrant-client>=1.7.0  # 升级到最新版本以兼容 Qdrant 服务器 latest 版本

# 任务队列
celery==5.3.4
redis==5.0.1

# LLM 调用
openai==1.3.5
dashscope==1.17.0  # 通义千问
httpx==0.25.1

# 数据处理
langchain==0.0.350
langchain-community==0.0.10
sentence-transformers>=2.7.0  # 本地 Embedding（升级到兼容新版本 huggingface_hub）
huggingface-hub>=0.20.0  # sentence-transformers 依赖
jieba==0.42.1  # 中文分词

# Markdown 处理
markdown==3.5.1
markdown-it-py==3.0.0

# 工具库
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyyaml==6.0.1
aiofiles==23.2.1

# 工具
python-jose[cryptography]==3.3.0  # JWT 认证
passlib[bcrypt]==1.7.4
psutil==5.9.6  # 内存监控（可选，用于获取内存使用信息）

